Move get_version_from_pyproject_toml func to utils
Move a slightly modified version of get_version_from_pyproject_toml
function to the utils module.
