+++ gvm/utils.py
@@ -19,6 +19,10 @@
 import re
 import warnings
 
+from pathlib import Path
+
+import toml
+
 from packaging.version import Version, InvalidVersion
 
 
@@ -39,6 +43,24 @@ def safe_version(version: str) -> str:
     except InvalidVersion:
         version = version.replace(' ', '.')
         return re.sub('[^A-Za-z0-9.]+', '-', version)
+
+
+def get_version_from_pyproject_toml() -> str:
+    """
+    Return the version information from the [tool.poetry] section of the
+    pyproject.toml file. The version may be in non standardized form.
+    """
+    path = Path(__file__)
+    pyproject_toml_path = path.parent.parent / 'pyproject.toml'
+
+    if not pyproject_toml_path.exists():
+        raise RuntimeError('pyproject.toml file not found.')
+
+    pyproject_toml = toml.loads(pyproject_toml_path.read_text())
+    if 'tool' in pyproject_toml and 'poetry' in pyproject_toml['tool']:
+        return pyproject_toml['tool']['poetry']['version']
+
+    raise RuntimeError('Version information not found in pyproject.toml file.')
 
 
 def get_version_string(version: tuple) -> str:
